Refactor generator tests to rely on a common base class for common fixtures.
