+++ tests/generators/test_ode_position_generator.py
@@ -1,34 +1,18 @@
-from typing import AnyStr, Dict
-
 import pytest
 import torch
 
 from crystal_diffusion.generators.ode_position_generator import (
     ExplodingVarianceODEPositionGenerator, ODESamplingParameters)
-from crystal_diffusion.models.score_networks.score_network import (
-    ScoreNetwork, ScoreNetworkParameters)
-from crystal_diffusion.namespace import NOISY_RELATIVE_COORDINATES
 from crystal_diffusion.samplers.variance_sampler import (
     ExplodingVarianceSampler, NoiseParameters)
-
-
-class FakeScoreNetwork(ScoreNetwork):
-    """A fake, smooth score network for the ODE solver."""
-
-    def _forward_unchecked(self, batch: Dict[AnyStr, torch.Tensor], conditional: bool = False) -> torch.Tensor:
-        return batch[NOISY_RELATIVE_COORDINATES]
+from tests.generators.conftest import BaseTestGenerator
 
 
 @pytest.mark.parametrize("total_time_steps", [2, 5, 10])
-@pytest.mark.parametrize("spatial_dimension", [2, 3])
-@pytest.mark.parametrize("number_of_atoms", [8])
 @pytest.mark.parametrize("sigma_min", [0.15])
 @pytest.mark.parametrize("record_samples", [False, True])
 @pytest.mark.parametrize("number_of_samples", [8])
-class TestExplodingVarianceODEPositionGenerator:
-    @pytest.fixture()
-    def sigma_normalized_score_network(self, spatial_dimension):
-        return FakeScoreNetwork(ScoreNetworkParameters(architecture='dummy', spatial_dimension=spatial_dimension))
+class TestExplodingVarianceODEPositionGenerator(BaseTestGenerator):
 
     @pytest.fixture()
     def noise_parameters(self, total_time_steps, sigma_min):
@@ -50,11 +34,6 @@ class TestExplodingVarianceODEPositionGenerator:
                                                           sigma_normalized_score_network=sigma_normalized_score_network)
 
         return generator
-
-    @pytest.fixture()
-    def unit_cell_sample(self, spatial_dimension, number_of_samples):
-        unit_cell_size = 10.
-        return torch.diag(torch.Tensor([unit_cell_size] * spatial_dimension)).repeat(number_of_samples, 1, 1)
 
     def test_get_exploding_variance_sigma(self, ode_generator, noise_parameters):
         times = ExplodingVarianceSampler._get_time_array(noise_parameters)
